requirements.txt
grpcio
grpcio-tools
python-jose
kafka-python
pymongo
protobuf
boto3
prometheus-client
locust